requirements.txt
azure-storage-queue
python-dotenv
